+++ lib/src/main/java/io/github/ascasso/garmin/livetrack/exception/LiveTrackException.java
@@ -1,0 +1,11 @@
+package io.github.ascasso.garmin.livetrack.exception;
+
+public class LiveTrackException extends RuntimeException {
+    public LiveTrackException(String message) {
+        super(message);
+    }
+
+    public LiveTrackException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}